requirements.txt
appdirs==1.4.0
click==6.7
Flask==0.12
google-api-python-client==1.6.2
httplib2==0.10.3
itsdangerous==0.24
Jinja2==2.9.5
MarkupSafe==0.23
oauth2client==4.0.0
olefile==0.44
packaging==16.8
Pillow==4.0.0
pyasn1==0.2.2
pyasn1-modules==0.0.8
pyparsing==2.1.10
rsa==3.4.2
uritemplate==3.0.0
Werkzeug==0.11.15
